fix up 0.1.0

+++ hooks/copy-tutorials.sh
@@ -1,9 +1,10 @@
-## Copy example notebooks to home
+## Create or clear Tutorials dir, copy examples and tutorials
+
 if [ ! -d "$HOME/notebooks/Tutorials" ]; then
   mkdir -p $HOME/notebooks/Tutorials
+elif [ -d "$HOME/notebooks/Tutorials" ]; then
+  rm -rf $HOME/notebooks/Tutorials/*
 fi
-
-rm -rf $HOME/notebooks/Tutorials/*
 
 mkdir $HOME/notebooks/Tutorials/pyCAP && cp -r /packages/pyCAP/examples/* $HOME/notebooks/Tutorials/pyCAP
 mkdir $HOME/notebooks/Tutorials/scdmsPyTools && cp -r /packages/scdmsPyTools/demo/* $HOME/notebooks/Tutorials/scdmsPyTools
